ABI Encoding for all combinations of arrays
